Flexbox positioning
This throws out the old absolute positioning method (which proves harmful because you don't know how clients are going to render the page at all). A problem I observed in Firefox was that my font was different (Roboto) than whatever the hell Chrome uses, which made the page look different because of the differences in pixel sizes of the fonts.